refactor: few components that I missed, types in types
